Fix serial number access, add lpcsdr_get_serial
There were a couple of bugs interacting (most egregiously, passing
completely the wrong length to libusb_get_string_descriptor_ascii) that
meant that the reported serial number was truncated to 8 chars. Fix
this.

Also add a separate interface to get the serial of an opened device
directly from the firmware, since the USB device info is gone by that
point.