lpcsdr_strerror_string: rename. fix case where buf is not populated
strerror was way too confusing a name. Make it clearer.

The call to lpcsdr_strerror_r returns a pointer to the string to use,
which may _not_ be the buffer the caller provided (e.g. if the string is
just a static string). Make sure we use the returned pointer to
construct our std::string