Rework of sampling rate / frequency API, prep for baseband work
This changes the library API for sample rate / frequency to,
essentially,

1) call lpcsdr_set_{sample_rate,frequency,sideband,conversion_mode}
   to set up the _requested_ settings
2) call lpcsdr_apply_changes to actually make the changes

In the baseband case, sample rate / freq / sideband settings interact in
ways that mean that intermediate states might be invalid (e.g. frequency
out of range because you just switched sideband mode). Options here are
either to have a big "set everything" function call (which I'm not a
huge fan of), or split up the configure vs. apply steps.

The corresponding getters for sample rate / frequency can fill in both
the requested value (what the user previously asked for) and, _after
changes are applied_, the actual hardware values including any
rounding/quantization etc.

Internally, we stash the generated ADC/tuner PLL configs as part of
lpcsdr_dev_handle for later reference. For the ADC PLL specifically,
we calculate the PLL settings at the time of apply (so we know that it's
in range and can determine what the actual hardware rate will be) but
don't actually set that on the hardware until we start streaming.