stream.c: implement baseband mode
Generalize some of the transfer/buffer size assumptions so they work for
both baseband & real mode; store some more metadata in the device about
this.

Add a spectrum-inverting version of unpack_raw_adc_data. Use this when
we're in baseband mode and using low-sideband tuning, so the unpacked
ADC data is always the uninverted version, and we don't need two
different downconverters for +/- Fs/4 mixing.

In BASEBAND mode, maintain a DSP downconverter instance & a temporary
ADC buffer in the device state for the duration of streaming.

Implement the BASEBAND case in convert_adc_blocks, which (a) unpacks ADC
data (with spectrum inversion if needed) to the temporary buffer, rather
than direct to a user buffer, then (b) calls the DSP downconverter to
shift that to baseband.