fs4_mix: properly initialize all the output
This is a stupid bug where it wasn't zeroing out the unused real/imag
parts of the output. It would "usually" work, if the ADC buffer happened
to be zeroed out (we malloc it, so it's unpredictable), but it would
often fail after changing the sampling rate because the new buffer would
reuse some old memory and have garbage, not zeros.

Manifested as extra noise, anywhere from "why do I get periodic bursts
with a little extra background noise" to "input is nothing but noise"