soapy: tweak tryApplyChanges et al to work better with CubicSDR
Add a PauseStreamGuard RAII-style helper so we can easily pause the
existing stream, change settings, then resume afterwards.

Fix up an activation/deactivation race that CubicSDR tended to provoke,
where it would call e.g.:

  -> activateStream
  worker thread begins to start up
  -> setSampleRate
  we try to deactivate the stream, but the worker's not running yet
  worker thread calls lpcsdr_start_streaming
  we call lpcsdr_set_sample_rate, but that fails because streaming is
now active again

We fix this by having a stop-flag that is set, and the worker thread
will bail out (from the buffer callback) when this is seen.