soapy: use the new gain API to fully implement the soapy gain API
This implements LNA/MIX/VGA gain stages as separate gain elements in the
SoapySDR API, and also overrides the default implementation of setGain
to use the gain curve.

CubicSDR doesn't work so well with the gain-curve API because it never
uses that variant of setGain, only ever looking at individual gain
elements. So also add a setting that optionally provides a fourth "ALL"
synthetic gain element that maps to the gain curve (i.e. changing the
gain of the "ALL" element implicitly changes the gains of LNA/MIX/VGA)