bandpass: improve filter selection logic vs Nyquist
Given the choice between:

a) a filter with an upper edge that is below Nyquist but not so close to
what we want, and

b) and a filter with an upper edge that is _above_ Nyquist, but where
the folded edge happens to be very close to what we want

we should probably not pick (b) as it will introduce a bunch of extra
noise from the span between the folded edge and the Nyquist frequency
(twice, because of folding)

e.g. for a desired filter of (1.75 - 5.25) with Nyquist=7, the old code
would pick a bandpass of (1.7 - 8.65) which is effectively (1.7 - 5.35)
after folding. The new code picks (1.7 - 5.69) which is probably better.