Add lpcsdr_{set,get}_undersampling_mode
This implements basic undersampling / bandpass sampling support. Clients
can set an undersampling factor N, for integer N >= 1

Usually (N=1) we tune the LO so that the RF signal we want appears at an
IF range of 0 to Fs/2, which is then captured at an ADC rate of Fs.

With N>1 we instead tune the RF signal to appear at higher frequencies
in the IF signal (e.g. N=2 => Fs/2 .. Fs; N=3 => Fs .. Fs*3/2) with a
matching tuner bandpass filter. The ADC still samples at a rate of Fs,
so the IF signal is above the Nyquist frequency, and we rely on aliasing
effects to capture one of the spectral replicas of the IF signal.

see Lyons section 2.3 pg 42 "Sampling bandpass signals"