Refine LIBUSB_ERROR_PIPE in response to tuner control transfers
Tuner control transfers will stall the control endpoint in response to
an I2C error; but they could also stall for other reasons.

When we see a stall in the context of a tuner control transfer, get the
current board status and check for the I2C-error flag. If it's set,
remap the generic libusb error to a more specific "I2C error" error
code.