lpcsdr__find_pll_parameters: allow nominally out of range VCO frequencies
If the requested frequency requires a VCO frequency that's nominally out
of range, but everything else is in range, then return those parameters
anyway.

For marginally out of range VCO frequencies there's a good chance that
the PLL will still lock anyway. If it doesn't lock, that's okay, we'll
return a PLL lock error at that point.