Fix timestamp discontinuity with particular combinations of settings
If the number of ADC samples per USB transfer (which depends on buffer
size) was not an exact multiple of the number of ADC samples per user
sample (which depends on decimation), then there would be periodic
1-sample discontinuities in the timestamp reported to user code, which
user code e.g. the soapy driver could interpret as dropped samples.

This happens because with those settings, there are some trailing
samples at the end of each transfer that have been passed to the DSP
chain and buffered, but which have not yet resulted in a user sample.
Those trailing samples will contribute to the first sample of the next
transfer. This means that the number of samples generated from a USB
transfer varies by +/- 1 sample.

The timestamp calculation would have the same behaviour, with +/- 1
sample variation in the difference in timestamp between successive user
buffers, but it didn't necessarily line up with what was happening
internally in the DSP code. Where it didn't line up, it would look to
user code like we had 1 extra sample, or dropped one sample.

To fix it, track the fractional "partial" user sample in terms of the
number of pending ADC samples, and use that to fix up the sample count
before computing the user timestamp. That is, pretend that the first
sample we're emitting from this transfer actually arrived at a time N
samples before the start-of-this-ADC-block, where N is the number of
buffered but unprocessed ADC samples left over from the previous
transfer.

Test settings to provoke it: 1MHz complex baseband, decimation=3 (divide
by 8), buffer_size=32768. This produces adc_samples_per_user_sample =
16, and USB transfers containing 77*6808=524216 samples which is not a
multiple of 16.